Refactor and add AbilityChart
