+++ iOS/Game/UI/MainPage/ShadowingListPage.swift
@@ -12,15 +12,13 @@ import Foundation
 
 private let context = GameContext.shared
 private let engine = SpeechEngine.shared
-private let activities = ["旅遊", "日常", "雜談", "戀愛", "論述", "敬語", "互動", "表達"]
 
-private let fontSize = screen.width * 12 / 320
 class ShadowingListPage: UIViewController {
     @IBOutlet weak var sentencesTableView: UITableView!
     @IBOutlet weak var timeline: TimelineView!
     @IBOutlet weak var topView: UIView!
     @IBOutlet weak var topBarView: TopBarView!
-    @IBOutlet weak var radarChartView: RadarChartView!
+    @IBOutlet weak var abilityChart: AbilityChart!
 
     var timelineSubviews: [String: UIView] = [:]
 
@@ -52,7 +50,7 @@ class ShadowingListPage: UIViewController {
 
         sentencesTableView.reloadData()
         timeline.viewWillAppear()
-        setChartData()
+        abilityChart.setChartData()
     }
 
     override func viewDidAppear(_ animated: Bool) {
@@ -116,76 +114,5 @@ extension ShadowingListPage: UITableViewDelegate {
         context.dataSetKey = allSentencesKeys[indexPath.row]
         context.loadLearningSentences(isShuffle: false)
         (UIApplication.getPresentedViewController() as? UIPageViewController)?.goToNextPage()
-    }
-}
-
-extension ShadowingListPage: IAxisValueFormatter {
-    func stringForValue(_ value: Double, axis: AxisBase?) -> String {
-        return activities[Int(value) % activities.count]
-    }
-}
-
-extension ShadowingListPage: ChartViewDelegate {
-    func setChartData() {
-        radarChartView.delegate = self
-
-        radarChartView.chartDescription?.enabled = false
-        radarChartView.webLineWidth = 0.5
-        radarChartView.innerWebLineWidth = 0.5
-        radarChartView.webColor = myGray.withAlphaComponent(0.7)
-        radarChartView.innerWebColor = myGray.withAlphaComponent(0.7)
-        radarChartView.webAlpha = 1
-
-        let xAxis = radarChartView.xAxis
-        xAxis.labelFont = MyFont.thin(ofSize: fontSize)
-        xAxis.xOffset = 0
-        xAxis.yOffset = 0
-        xAxis.valueFormatter = self
-        xAxis.labelTextColor = hashtagColor
-
-        let yAxis = radarChartView.yAxis
-        yAxis.labelFont = MyFont.thin(ofSize: fontSize)
-        yAxis.labelCount = 3
-        yAxis.axisMinimum = 0
-        yAxis.axisMaximum = 500
-        yAxis.drawLabelsEnabled = false
-
-        let l = radarChartView.legend
-        l.horizontalAlignment = .center
-        l.verticalAlignment = .bottom
-        l.orientation = .horizontal
-        l.drawInside = true
-        l.font = MyFont.thin(ofSize: fontSize)
-        l.xEntrySpace = 7
-        l.yEntrySpace = 5
-        l.textColor = .black
-        l.form = .none
-
-        radarChartView.animate(xAxisDuration: 1.4, yAxisDuration: 1.4, easingOption: .easeOutBack)
-
-        let tagScores = getTagScores()
-        let entries1 = activities.map { tag -> RadarChartDataEntry in
-            if let score = tagScores["#\(tag)"],
-               score > 10 {
-                return RadarChartDataEntry(value: Double(score))
-            }
-            return RadarChartDataEntry(value: 40)
-        }
-
-        let set1 = RadarChartDataSet(values: entries1, label: "")
-        set1.setColor(myOrange)
-        set1.fillColor = myOrange.withAlphaComponent(0.3)
-        set1.drawFilledEnabled = true
-        set1.fillAlpha = 0.5
-        set1.lineWidth = 1
-        set1.drawHighlightCircleEnabled = false
-        set1.setDrawHighlightIndicators(false)
-
-        let data = RadarChartData(dataSets: [set1])
-        data.setValueFont(MyFont.thin(ofSize: 8))
-        data.setDrawValues(false)
-        data.setValueTextColor(.black)
-
-        radarChartView.data = data
     }
 }
